Remove part that no longer seems active.
Is the "Talk to us directly" referenced on the readme still active? On the readme is says "Talk to us directly! Use this link to book time with a member of the product team to tell us more about your remote setup and experience so far." I did that but when I tried to log into the call the conference ID wasn't found. I also noticed that the person it said I was going to be talking with is no longer at Microsoft. If it is active and it just didn't work on this time then feel free to close this PR.