Create emote-ssh-stable-commands.md
Closes https://github.com/microsoft/vscode-remote-release/issues/7549
We got this request from an extension developer, in addition to just writing this doc we would have to commit to maintaining these functions. I think this is reasonable given how vital they are to the core functionality of the Remote-SSH extension. If this is carried through, comments should be added alongside the functions and the interface to alert future developers.