Init baseline configurations (https://github.com/microsoft/vscode-remote-release/issues/10248)
This is a start to an initiative aimed at collecting a set of
configurations to represent Remote - SSH compatibility against a wide
variety of inputs, distros, etc.

Tooling has been built internally, based on the dev container CLI,
to exercise the extension against this (growing) list of configurations.

Stay tuned into the following iteration plans for more progress on this
initiative.